+++ .eslintrc
@@ -1,0 +1,47 @@
+{
+  "extends": "airbnb-base",
+  "parser": "@babel/eslint-parser",
+  "rules": {
+    "arrow-parens": ["error", "always"],
+    "import/no-named-as-default": 0,
+    "import/prefer-default-export": 0,
+    "quotes": ["error", "single", { "avoidEscape": true, "allowTemplateLiterals": true }],
+    "comma-dangle": 0,
+    "object-curly-newline": ["error", {"ObjectPattern": { "multiline": true }}],
+    "eqeqeq": [1, "allow-null"],
+    "no-continue": 0,
+    "no-cond-assign": 1,
+    "no-constant-condition": 0,
+    "no-control-regex": 1,
+    "no-debugger": 1,
+    "no-dupe-keys": 1,
+    "no-ex-assign": 1,
+    "no-extra-boolean-cast": 1,
+    "no-func-assign": 1,
+    "no-regex-spaces": 1,
+    "no-unreachable": 1,
+    "no-fallthrough": 1,
+    "no-lone-blocks": 1,
+    "no-delete-var": 1,
+    "no-shadow": 1,
+    "no-shadow-restricted-names": 1,
+    "no-undef": 2,
+    "no-undef-init": 1,
+    "no-use-before-define": 0,
+    "no-unused-vars": [1, {"vars": "all", "args": "none"}],
+    "no-underscore-dangle": 0,
+    "no-restricted-syntax": ["error", "ForInStatement", "LabeledStatement", "WithStatement"],
+    "no-await-in-loop": "off",
+    "no-plusplus": "off"
+  },
+  "overrides": [
+    {
+      "files": [
+        "**/*.test.js"
+      ],
+      "env": {
+        "jest": true
+      }
+    }
+  ]
+}
