+++ __tests__/integration/tx-history.test.js
@@ -1,14 +1,16 @@
-import {TestUtils, WalletHelper} from "./test-utils-integration";
+import { TestUtils } from './utils/test-utils-integration';
+import { WalletHelper } from './utils/wallet-helper';
 
 describe('tx-history routes', () => {
   /** @type WalletHelper */
-  let wallet1, wallet2;
+  let wallet1;
+  let wallet2;
 
   /**
    * A map of transactions, where the key is the tx_id
    * @type {Record<string,unknown>}
    */
-  let fundTransactions = {};
+  const fundTransactions = {};
 
   beforeAll(async () => {
     try {
@@ -19,14 +21,12 @@ describe('tx-history routes', () => {
       // A wallet with 5 transactions containing 10, 20, 30, 40 and 50 HTR each
       wallet2 = new WalletHelper('txHistory2');
       await wallet2.start();
-      for (let amount = 10; amount < 60; amount = amount + 10) {
+      for (let amount = 10; amount < 60; amount += 10) {
         const fundTx = await wallet2.injectFunds(amount, 1);
         fundTransactions[fundTx.hash] = fundTx;
       }
-
-
     } catch (err) {
-      console.error(err.stack);
+      TestUtils.logError(err.stack);
     }
   });
 
@@ -35,20 +35,20 @@ describe('tx-history routes', () => {
     await wallet2.stop();
   });
 
-  it('should return an empty array of transactions for an empty wallet', async done => {
+  it('should return an empty array of transactions for an empty wallet', async (done) => {
     const balanceResult = await TestUtils.request
       .get('/wallet/tx-history')
-      .set({"x-wallet-id": wallet1.walletId});
+      .set({ 'x-wallet-id': wallet1.walletId });
 
     expect(balanceResult.status).toBe(200);
     expect(balanceResult.body).toHaveLength(0);
     done();
   });
 
-  it('should return transactions for a wallet', async done => {
+  it('should return transactions for a wallet', async (done) => {
     const balanceResult = await TestUtils.request
       .get('/wallet/tx-history')
-      .set({"x-wallet-id": wallet2.walletId});
+      .set({ 'x-wallet-id': wallet2.walletId });
 
     expect(balanceResult.status).toBe(200);
     const transactions = balanceResult.body;
@@ -57,7 +57,7 @@ describe('tx-history routes', () => {
       const transaction = transactions[txIndex];
       const fundTx = fundTransactions[transaction.tx_id];
 
-      // If the two first outputs and the first input are the same, it will be enough evidence for this test
+      // If the two first outputs and the first input are the same, it will be enough assertion
       expect(fundTx).toBeTruthy();
       expect(transaction.inputs[0].tx_id).toEqual(fundTx.inputs[0].tx_id);
       expect(transaction.outputs[0].value).toEqual(fundTx.outputs[0].value);
@@ -66,11 +66,11 @@ describe('tx-history routes', () => {
     done();
   });
 
-  it('should return transactions for a wallet with query limit', async done => {
+  it('should return transactions for a wallet with query limit', async (done) => {
     const balanceResult = await TestUtils.request
       .get('/wallet/tx-history')
-      .query({limit: 3})
-      .set({"x-wallet-id": wallet2.walletId});
+      .query({ limit: 3 })
+      .set({ 'x-wallet-id': wallet2.walletId });
 
     expect(balanceResult.status).toBe(200);
     const transactions = balanceResult.body;
@@ -79,10 +79,10 @@ describe('tx-history routes', () => {
       const transaction = transactions[txIndex];
       const fundTx = fundTransactions[transaction.tx_id];
 
-      // If we found the transaction it is evidence enough that it worked, since it is a hash of every other data there.
+      // If we found the transaction it is evidence enough that it worked,
+      // since it is a hash of every other data there.
       expect(fundTx).toBeTruthy();
     }
     done();
   });
-
 });
